sui_bridge::abi
Module eth_bridge_config

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddTokensWithSignaturesCall

Container type for all input parameters for the addTokensWithSignatures function with signature addTokensWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0x43025664

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

ChainIDCall

Container type for all input parameters for the chainID function with signature chainID() and selector 0xadc879e9

ChainIDReturn

Container type for all return fields from the chainID function with signature chainID() and selector 0xadc879e9

CommitteeCall

Container type for all input parameters for the committee function with signature committee() and selector 0xd864e740

CommitteeReturn

Container type for all return fields from the committee function with signature committee() and selector 0xd864e740

ContractUpgradedFilter

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

EthBridgeConfig

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address,uint8,address[],uint64[],uint8[]) and selector 0xdcb09117

InitializedFilter

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

IsChainSupportedCall

Container type for all input parameters for the isChainSupported function with signature isChainSupported(uint8) and selector 0x73209533

IsChainSupportedReturn

Container type for all return fields from the isChainSupported function with signature isChainSupported(uint8) and selector 0x73209533

IsTokenSupportedCall

Container type for all input parameters for the isTokenSupported function with signature isTokenSupported(uint8) and selector 0x438cb46a

IsTokenSupportedReturn

Container type for all return fields from the isTokenSupported function with signature isTokenSupported(uint8) and selector 0x438cb46a

Message

Message(uint8,uint8,uint64,uint8,bytes)

NoncesCall

Container type for all input parameters for the nonces function with signature nonces(uint8) and selector 0x5035bda2

NoncesReturn

Container type for all return fields from the nonces function with signature nonces(uint8) and selector 0x5035bda2

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ReentrancyGuardReentrantCall

Custom Error type ReentrancyGuardReentrantCall with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5

SupportedChainsCall

Container type for all input parameters for the supportedChains function with signature supportedChains(uint8) and selector 0x7b099ce5

SupportedChainsReturn

Container type for all return fields from the supportedChains function with signature supportedChains(uint8) and selector 0x7b099ce5

SupportedTokensCall

Container type for all input parameters for the supportedTokens function with signature supportedTokens(uint8) and selector 0x3f572d36

SupportedTokensReturn

Container type for all return fields from the supportedTokens function with signature supportedTokens(uint8) and selector 0x3f572d36

TokenAddedFilter

TokenAddressOfCall

Container type for all input parameters for the tokenAddressOf function with signature tokenAddressOf(uint8) and selector 0xe5324889

TokenAddressOfReturn

Container type for all return fields from the tokenAddressOf function with signature tokenAddressOf(uint8) and selector 0xe5324889

TokenPriceOfCall

Container type for all input parameters for the tokenPriceOf function with signature tokenPriceOf(uint8) and selector 0xdfc3db3d

TokenPriceOfReturn

Container type for all return fields from the tokenPriceOf function with signature tokenPriceOf(uint8) and selector 0xdfc3db3d

TokenPriceUpdatedFilter

TokenPriceUpdatedV2Filter

TokenPricesCall

Container type for all input parameters for the tokenPrices function with signature tokenPrices(uint8) and selector 0xdeceafa7

TokenPricesReturn

Container type for all return fields from the tokenPrices function with signature tokenPrices(uint8) and selector 0xdeceafa7

TokenSuiDecimalOfCall

Container type for all input parameters for the tokenSuiDecimalOf function with signature tokenSuiDecimalOf(uint8) and selector 0x71ceee4f

TokenSuiDecimalOfReturn

Container type for all return fields from the tokenSuiDecimalOf function with signature tokenSuiDecimalOf(uint8) and selector 0x71ceee4f

TokensAddedV2Filter

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UpdateTokenPriceWithSignaturesCall

Container type for all input parameters for the updateTokenPriceWithSignatures function with signature updateTokenPriceWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xbfb5d846

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradeWithSignaturesCall

Container type for all input parameters for the upgradeWithSignatures function with signature upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xfa72a6d0

UpgradedFilter

Enums

EthBridgeConfigCalls

Container type for all of the contract’s call

EthBridgeConfigErrors

Container type for all of the contract’s custom errors

EthBridgeConfigEvents

Container type for all of the contract’s events

Statics

ETHBRIDGECONFIG_ABI

The parsed JSON ABI of the contract.