Module sui_bridge::abi::eth_bridge_vault

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

EthBridgeVault

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

TransferERC20Call

Container type for all input parameters for the transferERC20 function with signature transferERC20(address,address,uint256) and selector 0x9db5dbe4

TransferETHCall

Container type for all input parameters for the transferETH function with signature transferETH(address,uint256) and selector 0x7b1a4909

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

WethCall

Container type for all input parameters for the wETH function with signature wETH() and selector 0xf2428621

WethReturn

Container type for all return fields from the wETH function with signature wETH() and selector 0xf2428621

Enums

EthBridgeVaultCalls

Container type for all of the contract’s call

EthBridgeVaultErrors

Container type for all of the contract’s custom errors

Statics

ETHBRIDGEVAULT_ABI

The parsed JSON ABI of the contract.