sui_bridge::abi::eth_bridge_vault

Module EthBridgeVault

Generated by the following Solidity interface…

interface EthBridgeVault {
    error AddressEmptyCode(address target);
    error AddressInsufficientBalance(address account);
    error FailedInnerCall();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error ReentrancyGuardReentrantCall();
    error SafeERC20FailedOperation(address token);

    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);

    constructor(address _wETH);

    receive() external payable;

    function owner() external view returns (address);
    function renounceOwnership() external;
    function transferERC20(address tokenAddress, address recipientAddress, uint256 amount) external;
    function transferETH(address payable recipientAddress, uint256 amount) external;
    function transferOwnership(address newOwner) external;
    function wETH() external view returns (address);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_wETH",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "receive",
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferERC20",
    "inputs": [
      {
        "name": "tokenAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "recipientAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferETH",
    "inputs": [
      {
        "name": "recipientAddress",
        "type": "address",
        "internalType": "address payable"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "wETH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IWETH9"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AddressInsufficientBalance",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "FailedInnerCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ReentrancyGuardReentrantCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "SafeERC20FailedOperation",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ]
  }
]

Structs

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AddressInsufficientBalance

Custom error with signature AddressInsufficientBalance(address) and selector 0xcd786059.

EthBridgeVaultInstance

A EthBridgeVault instance.

FailedInnerCall

Custom error with signature FailedInnerCall() and selector 0x1425ea42.

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7.

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

ReentrancyGuardReentrantCall

Custom error with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5.

SafeERC20FailedOperation

Custom error with signature SafeERC20FailedOperation(address) and selector 0x5274afe7.

constructorCall

Constructor`.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

transferERC20Call

Function with signature transferERC20(address,address,uint256) and selector 0x9db5dbe4.

transferERC20Return

Container type for the return parameters of the transferERC20(address,address,uint256) function.

transferETHCall

Function with signature transferETH(address,uint256) and selector 0x7b1a4909.

transferETHReturn

Container type for the return parameters of the transferETH(address,uint256) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

wETHCall

Function with signature wETH() and selector 0xf2428621.

wETHReturn

Container type for the return parameters of the wETH() function.

Enums

EthBridgeVaultCalls

Container for all the EthBridgeVault function calls.

EthBridgeVaultErrors

Container for all the EthBridgeVault custom errors.

EthBridgeVaultEvents

Container for all the EthBridgeVault events.

Functions

new

Creates a new wrapper around an on-chain EthBridgeVault contract instance.