Module sui_bridge::abi::eth_sui_bridge

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AdjustDecimalsForErc20Call

Container type for all input parameters for the adjustDecimalsForErc20 function with signature adjustDecimalsForErc20(uint8,uint64,uint8) and selector 0xd280b085

AdjustDecimalsForErc20Return

Container type for all return fields from the adjustDecimalsForErc20 function with signature adjustDecimalsForErc20(uint8,uint64,uint8) and selector 0xd280b085

AdjustDecimalsForSuiTokenCall

Container type for all input parameters for the adjustDecimalsForSuiToken function with signature adjustDecimalsForSuiToken(uint8,uint256,uint8) and selector 0x52d07066

AdjustDecimalsForSuiTokenReturn

Container type for all return fields from the adjustDecimalsForSuiToken function with signature adjustDecimalsForSuiToken(uint8,uint256,uint8) and selector 0x52d07066

AdminChangedFilter

BeaconUpgradedFilter

BridgeETHToSuiCall

Container type for all input parameters for the bridgeETHToSui function with signature bridgeETHToSui(bytes,uint8) and selector 0x11571cc2

BridgeToSuiCall

Container type for all input parameters for the bridgeToSui function with signature bridgeToSui(uint8,uint256,bytes,uint8) and selector 0x780aad0a

BridgeUpgradeStakeRequiredCall

Container type for all input parameters for the BRIDGE_UPGRADE_STAKE_REQUIRED function with signature BRIDGE_UPGRADE_STAKE_REQUIRED() and selector 0x402911ed

BridgeUpgradeStakeRequiredReturn

Container type for all return fields from the BRIDGE_UPGRADE_STAKE_REQUIRED function with signature BRIDGE_UPGRADE_STAKE_REQUIRED() and selector 0x402911ed

ChainIdCall

Container type for all input parameters for the chainId function with signature chainId() and selector 0x9a8a0592

ChainIdReturn

Container type for all return fields from the chainId function with signature chainId() and selector 0x9a8a0592

CommitteeCall

Container type for all input parameters for the committee function with signature committee() and selector 0xd864e740

CommitteeReturn

Container type for all return fields from the committee function with signature committee() and selector 0xd864e740

EthSuiBridge

ExecuteEmergencyOpWithSignaturesCall

Container type for all input parameters for the executeEmergencyOpWithSignatures function with signature executeEmergencyOpWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xa6f740f6

FreezingStakeRequiredCall

Container type for all input parameters for the FREEZING_STAKE_REQUIRED function with signature FREEZING_STAKE_REQUIRED() and selector 0x86375d6e

FreezingStakeRequiredReturn

Container type for all return fields from the FREEZING_STAKE_REQUIRED function with signature FREEZING_STAKE_REQUIRED() and selector 0x86375d6e

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address[],address,address,address,uint8) and selector 0x43ef9307

InitializedFilter

Message

Message(uint8,uint8,uint64,uint8,bytes)

MessageProcessedCall

Container type for all input parameters for the messageProcessed function with signature messageProcessed(uint64) and selector 0x20dd94b6

MessageProcessedReturn

Container type for all return fields from the messageProcessed function with signature messageProcessed(uint64) and selector 0x20dd94b6

NoncesCall

Container type for all input parameters for the nonces function with signature nonces(uint8) and selector 0x5035bda2

NoncesReturn

Container type for all return fields from the nonces function with signature nonces(uint8) and selector 0x5035bda2

PausedCall

Container type for all input parameters for the paused function with signature paused() and selector 0x5c975abb

PausedFilter

PausedReturn

Container type for all return fields from the paused function with signature paused() and selector 0x5c975abb

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

RequiredApprovalStakeCall

Container type for all input parameters for the requiredApprovalStake function with signature requiredApprovalStake(uint8) and selector 0xdcaffc0a

RequiredApprovalStakeReturn

Container type for all return fields from the requiredApprovalStake function with signature requiredApprovalStake(uint8) and selector 0xdcaffc0a

SupportedTokensCall

Container type for all input parameters for the supportedTokens function with signature supportedTokens(uint8) and selector 0x3f572d36

SupportedTokensReturn

Container type for all return fields from the supportedTokens function with signature supportedTokens(uint8) and selector 0x3f572d36

TokensBridgedToSuiFilter

TransferStakeRequiredCall

Container type for all input parameters for the TRANSFER_STAKE_REQUIRED function with signature TRANSFER_STAKE_REQUIRED() and selector 0xf2bd71bc

TransferStakeRequiredReturn

Container type for all return fields from the TRANSFER_STAKE_REQUIRED function with signature TRANSFER_STAKE_REQUIRED() and selector 0xf2bd71bc

TransferTokensWithSignaturesCall

Container type for all input parameters for the transferTokensWithSignatures function with signature transferTokensWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0x45134db3

UnfreezingStakeRequiredCall

Container type for all input parameters for the UNFREEZING_STAKE_REQUIRED function with signature UNFREEZING_STAKE_REQUIRED() and selector 0x92596ee5

UnfreezingStakeRequiredReturn

Container type for all return fields from the UNFREEZING_STAKE_REQUIRED function with signature UNFREEZING_STAKE_REQUIRED() and selector 0x92596ee5

UnpausedFilter

UpgradeBridgeWithSignaturesCall

Container type for all input parameters for the upgradeBridgeWithSignatures function with signature upgradeBridgeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0x37cacf5c

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradeToCall

Container type for all input parameters for the upgradeTo function with signature upgradeTo(address) and selector 0x3659cfe6

UpgradedFilter

VaultCall

Container type for all input parameters for the vault function with signature vault() and selector 0xfbfa77cf

VaultReturn

Container type for all return fields from the vault function with signature vault() and selector 0xfbfa77cf

Weth9Call

Container type for all input parameters for the weth9 function with signature weth9() and selector 0x50879c1c

Weth9Return

Container type for all return fields from the weth9 function with signature weth9() and selector 0x50879c1c

Enums

EthSuiBridgeCalls

Container type for all of the contract’s call

EthSuiBridgeEvents

Container type for all of the contract’s events

Statics

ETHSUIBRIDGE_ABI

The parsed JSON ABI of the contract.