sui_bridge::abi::eth_sui_bridge

Module EthSuiBridge

Generated by the following Solidity interface…

library BridgeUtils {
    struct Message {
        uint8 messageType;
        uint8 version;
        uint64 nonce;
        uint8 chainID;
        bytes payload;
    }
}

interface EthSuiBridge {
    error AddressEmptyCode(address target);
    error AddressInsufficientBalance(address account);
    error ERC1967InvalidImplementation(address implementation);
    error ERC1967NonPayable();
    error EnforcedPause();
    error ExpectedPause();
    error FailedInnerCall();
    error InvalidInitialization();
    error NotInitializing();
    error ReentrancyGuardReentrantCall();
    error SafeERC20FailedOperation(address token);
    error UUPSUnauthorizedCallContext();
    error UUPSUnsupportedProxiableUUID(bytes32 slot);

    event ContractUpgraded(uint256 nonce, address proxy, address implementation);
    event EmergencyOperation(uint64 nonce, bool paused);
    event Initialized(uint64 version);
    event Paused(address account);
    event TokensClaimed(uint8 indexed sourceChainID, uint64 indexed nonce, uint8 indexed destinationChainID, uint8 tokenID, uint256 erc20AdjustedAmount, bytes senderAddress, address recipientAddress);
    event TokensDeposited(uint8 indexed sourceChainID, uint64 indexed nonce, uint8 indexed destinationChainID, uint8 tokenID, uint64 suiAdjustedAmount, address senderAddress, bytes recipientAddress);
    event TokensDepositedV2(uint8 indexed sourceChainID, uint64 indexed nonce, uint8 indexed destinationChainID, uint8 tokenID, uint64 suiAdjustedAmount, address senderAddress, bytes recipientAddress, uint256 timestampSeconds);
    event Unpaused(address account);
    event Upgraded(address indexed implementation);

    function UPGRADE_INTERFACE_VERSION() external view returns (string memory);
    function bridgeERC20(uint8 tokenID, uint256 amount, bytes memory recipientAddress, uint8 destinationChainID) external;
    function bridgeERC20V2(uint8 tokenID, uint256 amount, bytes memory recipientAddress, uint8 destinationChainID) external;
    function bridgeETH(bytes memory recipientAddress, uint8 destinationChainID) external payable;
    function bridgeETHV2(bytes memory recipientAddress, uint8 destinationChainID) external payable;
    function committee() external view returns (address);
    function executeEmergencyOpWithSignatures(bytes[] memory signatures, BridgeUtils.Message memory message) external;
    function initialize(address _committee, address _vault, address _limiter) external;
    function isTransferProcessed(uint64 nonce) external view returns (bool isProcessed);
    function limiter() external view returns (address);
    function nonces(uint8 messageType) external view returns (uint64 nonce);
    function paused() external view returns (bool);
    function proxiableUUID() external view returns (bytes32);
    function transferBridgedTokensWithSignatures(bytes[] memory signatures, BridgeUtils.Message memory message) external;
    function transferBridgedTokensWithSignaturesV2(bytes[] memory signatures, BridgeUtils.Message memory message) external;
    function upgradeToAndCall(address newImplementation, bytes memory data) external payable;
    function upgradeWithSignatures(bytes[] memory signatures, BridgeUtils.Message memory message) external;
    function vault() external view returns (address);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "function",
    "name": "UPGRADE_INTERFACE_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "bridgeERC20",
    "inputs": [
      {
        "name": "tokenID",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "recipientAddress",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bridgeERC20V2",
    "inputs": [
      {
        "name": "tokenID",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "recipientAddress",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bridgeETH",
    "inputs": [
      {
        "name": "recipientAddress",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "bridgeETHV2",
    "inputs": [
      {
        "name": "recipientAddress",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "committee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IBridgeCommittee"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "executeEmergencyOpWithSignatures",
    "inputs": [
      {
        "name": "signatures",
        "type": "bytes[]",
        "internalType": "bytes[]"
      },
      {
        "name": "message",
        "type": "tuple",
        "internalType": "struct BridgeUtils.Message",
        "components": [
          {
            "name": "messageType",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "version",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "nonce",
            "type": "uint64",
            "internalType": "uint64"
          },
          {
            "name": "chainID",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "payload",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "_committee",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_vault",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_limiter",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "isTransferProcessed",
    "inputs": [
      {
        "name": "nonce",
        "type": "uint64",
        "internalType": "uint64"
      }
    ],
    "outputs": [
      {
        "name": "isProcessed",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "limiter",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IBridgeLimiter"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "nonces",
    "inputs": [
      {
        "name": "messageType",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [
      {
        "name": "nonce",
        "type": "uint64",
        "internalType": "uint64"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "paused",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "proxiableUUID",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "transferBridgedTokensWithSignatures",
    "inputs": [
      {
        "name": "signatures",
        "type": "bytes[]",
        "internalType": "bytes[]"
      },
      {
        "name": "message",
        "type": "tuple",
        "internalType": "struct BridgeUtils.Message",
        "components": [
          {
            "name": "messageType",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "version",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "nonce",
            "type": "uint64",
            "internalType": "uint64"
          },
          {
            "name": "chainID",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "payload",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferBridgedTokensWithSignaturesV2",
    "inputs": [
      {
        "name": "signatures",
        "type": "bytes[]",
        "internalType": "bytes[]"
      },
      {
        "name": "message",
        "type": "tuple",
        "internalType": "struct BridgeUtils.Message",
        "components": [
          {
            "name": "messageType",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "version",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "nonce",
            "type": "uint64",
            "internalType": "uint64"
          },
          {
            "name": "chainID",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "payload",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "upgradeToAndCall",
    "inputs": [
      {
        "name": "newImplementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "upgradeWithSignatures",
    "inputs": [
      {
        "name": "signatures",
        "type": "bytes[]",
        "internalType": "bytes[]"
      },
      {
        "name": "message",
        "type": "tuple",
        "internalType": "struct BridgeUtils.Message",
        "components": [
          {
            "name": "messageType",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "version",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "nonce",
            "type": "uint64",
            "internalType": "uint64"
          },
          {
            "name": "chainID",
            "type": "uint8",
            "internalType": "uint8"
          },
          {
            "name": "payload",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "vault",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IBridgeVault"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "ContractUpgraded",
    "inputs": [
      {
        "name": "nonce",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "proxy",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "implementation",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "EmergencyOperation",
    "inputs": [
      {
        "name": "nonce",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      },
      {
        "name": "paused",
        "type": "bool",
        "indexed": false,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Paused",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TokensClaimed",
    "inputs": [
      {
        "name": "sourceChainID",
        "type": "uint8",
        "indexed": true,
        "internalType": "uint8"
      },
      {
        "name": "nonce",
        "type": "uint64",
        "indexed": true,
        "internalType": "uint64"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "indexed": true,
        "internalType": "uint8"
      },
      {
        "name": "tokenID",
        "type": "uint8",
        "indexed": false,
        "internalType": "uint8"
      },
      {
        "name": "erc20AdjustedAmount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "senderAddress",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      },
      {
        "name": "recipientAddress",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TokensDeposited",
    "inputs": [
      {
        "name": "sourceChainID",
        "type": "uint8",
        "indexed": true,
        "internalType": "uint8"
      },
      {
        "name": "nonce",
        "type": "uint64",
        "indexed": true,
        "internalType": "uint64"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "indexed": true,
        "internalType": "uint8"
      },
      {
        "name": "tokenID",
        "type": "uint8",
        "indexed": false,
        "internalType": "uint8"
      },
      {
        "name": "suiAdjustedAmount",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      },
      {
        "name": "senderAddress",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "recipientAddress",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TokensDepositedV2",
    "inputs": [
      {
        "name": "sourceChainID",
        "type": "uint8",
        "indexed": true,
        "internalType": "uint8"
      },
      {
        "name": "nonce",
        "type": "uint64",
        "indexed": true,
        "internalType": "uint64"
      },
      {
        "name": "destinationChainID",
        "type": "uint8",
        "indexed": true,
        "internalType": "uint8"
      },
      {
        "name": "tokenID",
        "type": "uint8",
        "indexed": false,
        "internalType": "uint8"
      },
      {
        "name": "suiAdjustedAmount",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      },
      {
        "name": "senderAddress",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "recipientAddress",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      },
      {
        "name": "timestampSeconds",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Unpaused",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgraded",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AddressInsufficientBalance",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967InvalidImplementation",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967NonPayable",
    "inputs": []
  },
  {
    "type": "error",
    "name": "EnforcedPause",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ExpectedPause",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedInnerCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ReentrancyGuardReentrantCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "SafeERC20FailedOperation",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "UUPSUnauthorizedCallContext",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnsupportedProxiableUUID",
    "inputs": [
      {
        "name": "slot",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  }
]

Structs

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AddressInsufficientBalance

Custom error with signature AddressInsufficientBalance(address) and selector 0xcd786059.

ContractUpgraded

Event with signature ContractUpgraded(uint256,address,address) and selector 0xdc69b57038334451ee12fd1742228917cea7f40dbd33cda5162e7e5754acee1c.

ERC1967InvalidImplementation

Custom error with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3.

ERC1967NonPayable

Custom error with signature ERC1967NonPayable() and selector 0xb398979f.

EmergencyOperation

Event with signature EmergencyOperation(uint64,bool) and selector 0x0838fecaac9057733ed7f9bb6e8ecfacf0d9c68d2f9b06e84f8c7f6098814bfb.

EnforcedPause

Custom error with signature EnforcedPause() and selector 0xd93c0665.

EthSuiBridgeInstance

A EthSuiBridge instance.

ExpectedPause

Custom error with signature ExpectedPause() and selector 0x8dfc202b.

FailedInnerCall

Custom error with signature FailedInnerCall() and selector 0x1425ea42.

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2.

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9.

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8.

Paused

Event with signature Paused(address) and selector 0x62e78cea01bee320cd4e420270b5ea74000d11b0c9f74754ebdbfc544b05a258.

ReentrancyGuardReentrantCall

Custom error with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5.

SafeERC20FailedOperation

Custom error with signature SafeERC20FailedOperation(address) and selector 0x5274afe7.

TokensClaimed

Event with signature TokensClaimed(uint8,uint64,uint8,uint8,uint256,bytes,address) and selector 0x933e8377dca7a8cf67d2bf865c4d8c1c45347815760900f9d2eb5655a06943af.

TokensDeposited

Event with signature TokensDeposited(uint8,uint64,uint8,uint8,uint64,address,bytes) and selector 0xa0f1d54820817ede8517e70a3d0a9197c015471c5360d2119b759f0359858ce6.

TokensDepositedV2

Event with signature TokensDepositedV2(uint8,uint64,uint8,uint8,uint64,address,bytes,uint256) and selector 0x70ef4a8b6f5065985fc424842df9fc218ad4c4015e85bb4158770e34abbcd82c.

UPGRADE_INTERFACE_VERSIONCall

Function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc.

UPGRADE_INTERFACE_VERSIONReturn

Container type for the return parameters of the UPGRADE_INTERFACE_VERSION() function.

UUPSUnauthorizedCallContext

Custom error with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba.

UUPSUnsupportedProxiableUUID

Custom error with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4.

Unpaused

Event with signature Unpaused(address) and selector 0x5db9ee0a495bf2e6ff9c91a7834c1ba4fdd244a5e8aa4e537bd38aeae4b073aa.

Upgraded

Event with signature Upgraded(address) and selector 0xbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b.

bridgeERC20Call

Function with signature bridgeERC20(uint8,uint256,bytes,uint8) and selector 0x212afaff.

bridgeERC20Return

Container type for the return parameters of the bridgeERC20(uint8,uint256,bytes,uint8) function.

bridgeERC20V2Call

Function with signature bridgeERC20V2(uint8,uint256,bytes,uint8) and selector 0x0bdba58b.

bridgeERC20V2Return

Container type for the return parameters of the bridgeERC20V2(uint8,uint256,bytes,uint8) function.

bridgeETHCall

Function with signature bridgeETH(bytes,uint8) and selector 0x9449ebd2.

bridgeETHReturn

Container type for the return parameters of the bridgeETH(bytes,uint8) function.

bridgeETHV2Call

Function with signature bridgeETHV2(bytes,uint8) and selector 0xb4b7068c.

bridgeETHV2Return

Container type for the return parameters of the bridgeETHV2(bytes,uint8) function.

committeeCall

Function with signature committee() and selector 0xd864e740.

committeeReturn

Container type for the return parameters of the committee() function.

executeEmergencyOpWithSignaturesCall

Function with signature executeEmergencyOpWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xa6f740f6.

executeEmergencyOpWithSignaturesReturn

Container type for the return parameters of the executeEmergencyOpWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) function.

initializeCall

Function with signature initialize(address,address,address) and selector 0xc0c53b8b.

initializeReturn

Container type for the return parameters of the initialize(address,address,address) function.

isTransferProcessedCall

Function with signature isTransferProcessed(uint64) and selector 0x5458ea9e.

isTransferProcessedReturn

Container type for the return parameters of the isTransferProcessed(uint64) function.

limiterCall

Function with signature limiter() and selector 0x74b87f67.

limiterReturn

Container type for the return parameters of the limiter() function.

noncesCall

Function with signature nonces(uint8) and selector 0x5035bda2.

noncesReturn

Container type for the return parameters of the nonces(uint8) function.

pausedCall

Function with signature paused() and selector 0x5c975abb.

pausedReturn

Container type for the return parameters of the paused() function.

proxiableUUIDCall

Function with signature proxiableUUID() and selector 0x52d1902d.

proxiableUUIDReturn

Container type for the return parameters of the proxiableUUID() function.

transferBridgedTokensWithSignaturesCall

Function with signature transferBridgedTokensWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xbeb0d55c.

transferBridgedTokensWithSignaturesReturn

Container type for the return parameters of the transferBridgedTokensWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) function.

transferBridgedTokensWithSignaturesV2Call

Function with signature transferBridgedTokensWithSignaturesV2(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0x9c3ecefb.

transferBridgedTokensWithSignaturesV2Return

Container type for the return parameters of the transferBridgedTokensWithSignaturesV2(bytes[],(uint8,uint8,uint64,uint8,bytes)) function.

upgradeToAndCallCall

Function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286.

upgradeToAndCallReturn

Container type for the return parameters of the upgradeToAndCall(address,bytes) function.

upgradeWithSignaturesCall

Function with signature upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xfa72a6d0.

upgradeWithSignaturesReturn

Container type for the return parameters of the upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) function.

vaultCall

Function with signature vault() and selector 0xfbfa77cf.

vaultReturn

Container type for the return parameters of the vault() function.

Enums

EthSuiBridgeCalls

Container for all the EthSuiBridge function calls.

EthSuiBridgeErrors

Container for all the EthSuiBridge custom errors.

EthSuiBridgeEvents

Container for all the EthSuiBridge events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain EthSuiBridge contract instance.