sui_bridge::abi::eth_sui_bridge::EthSuiBridge

Function deploy

pub fn deploy<P: Provider<N>, N: Network>(
    __provider: P,
) -> impl Future<Output = Result<EthSuiBridgeInstance<P, N>>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.