sui_bridge::abi::eth_sui_bridge
Struct EthSuiBridge

pub struct EthSuiBridge<M>(/* private fields */);

Implementations

impl<M: Middleware> EthSuiBridge<M>

pub fn new<T: Into<Address>>(address: T, client: Arc<M>) -> Self

Creates a new contract instance with the specified ethers client at address. The contract derefs to a ethers::Contract object.

pub fn upgrade_interface_version(&self) -> ContractCall<M, String>

Calls the contract’s UPGRADE_INTERFACE_VERSION (0xad3cb1cc) function

pub fn bridge_erc20( &self, token_id: u8, amount: U256, recipient_address: Bytes, destination_chain_id: u8, ) -> ContractCall<M, ()>

Calls the contract’s bridgeERC20 (0x212afaff) function

pub fn bridge_eth( &self, recipient_address: Bytes, destination_chain_id: u8, ) -> ContractCall<M, ()>

Calls the contract’s bridgeETH (0x9449ebd2) function

pub fn committee(&self) -> ContractCall<M, Address>

Calls the contract’s committee (0xd864e740) function

pub fn execute_emergency_op_with_signatures( &self, signatures: Vec<Bytes>, message: Message, ) -> ContractCall<M, ()>

Calls the contract’s executeEmergencyOpWithSignatures (0xa6f740f6) function

pub fn initialize( &self, committee: Address, vault: Address, limiter: Address, ) -> ContractCall<M, ()>

Calls the contract’s initialize (0xc0c53b8b) function

pub fn is_transfer_processed(&self, nonce: u64) -> ContractCall<M, bool>

Calls the contract’s isTransferProcessed (0x5458ea9e) function

pub fn limiter(&self) -> ContractCall<M, Address>

Calls the contract’s limiter (0x74b87f67) function

pub fn nonces(&self, message_type: u8) -> ContractCall<M, u64>

Calls the contract’s nonces (0x5035bda2) function

pub fn paused(&self) -> ContractCall<M, bool>

Calls the contract’s paused (0x5c975abb) function

pub fn proxiable_uuid(&self) -> ContractCall<M, [u8; 32]>

Calls the contract’s proxiableUUID (0x52d1902d) function

pub fn transfer_bridged_tokens_with_signatures( &self, signatures: Vec<Bytes>, message: Message, ) -> ContractCall<M, ()>

Calls the contract’s transferBridgedTokensWithSignatures (0xbeb0d55c) function

pub fn upgrade_to_and_call( &self, new_implementation: Address, data: Bytes, ) -> ContractCall<M, ()>

Calls the contract’s upgradeToAndCall (0x4f1ef286) function

pub fn upgrade_with_signatures( &self, signatures: Vec<Bytes>, message: Message, ) -> ContractCall<M, ()>

Calls the contract’s upgradeWithSignatures (0xfa72a6d0) function

pub fn vault(&self) -> ContractCall<M, Address>

Calls the contract’s vault (0xfbfa77cf) function

pub fn contract_upgraded_filter( &self, ) -> Event<Arc<M>, M, ContractUpgradedFilter>

Gets the contract’s ContractUpgraded event

pub fn emergency_operation_filter( &self, ) -> Event<Arc<M>, M, EmergencyOperationFilter>

Gets the contract’s EmergencyOperation event

pub fn initialized_filter(&self) -> Event<Arc<M>, M, InitializedFilter>

Gets the contract’s Initialized event

pub fn paused_filter(&self) -> Event<Arc<M>, M, PausedFilter>

Gets the contract’s Paused event

pub fn tokens_claimed_filter(&self) -> Event<Arc<M>, M, TokensClaimedFilter>

Gets the contract’s TokensClaimed event

pub fn tokens_deposited_filter(&self) -> Event<Arc<M>, M, TokensDepositedFilter>

Gets the contract’s TokensDeposited event

pub fn unpaused_filter(&self) -> Event<Arc<M>, M, UnpausedFilter>

Gets the contract’s Unpaused event

pub fn upgraded_filter(&self) -> Event<Arc<M>, M, UpgradedFilter>

Gets the contract’s Upgraded event

pub fn events(&self) -> Event<Arc<M>, M, EthSuiBridgeEvents>

Returns an Event builder for all the events of this contract.