Module sui_bridge::sui_bridge_watchdog

The BridgeWatchDog module is responsible for monitoring the health of the bridge by periodically running various observables and reporting the results.

Modules

eth_bridge_status

The EthBridgeStatus observable monitors whether the Eth Bridge is paused.

eth_vault_balance

metrics

sui_bridge_status

The SuiBridgeStatus observable monitors whether the Sui Bridge is paused.

total_supplies

The SuiBridgeStatus observable monitors whether the Sui Bridge is paused.

Structs

BridgeWatchDog

Traits

Observable