sui_bridge::utils

Function get_committee_voting_power_by_name

pub async fn get_committee_voting_power_by_name(
    bridge_committee: &Arc<BridgeCommittee>,
    system_state: &SuiSystemStateSummary,
) -> BTreeMap<String, StakeUnit>

Return a mappping from validator name to their bridge voting power. If a validator is not in the Sui committee, we will use its base URL as the name.