Crate sui_bridge

Modules

abi

action_executor

BridgeActionExecutor receives BridgeActions (from BridgeOrchestrator), collects bridge authority signatures and submit signatures on chain.

client

config

crypto

error

eth_client

eth_syncer

The EthSyncer module is responsible for synchronizing Events emitted on Ethereum blockchain from concerned contracts. Each contract is associated with a start block number, and the syncer will only query from that block number onwards. The syncer also keeps track of the last finalized block on Ethereum and will only query for events up to that block number.

events

This file contains the definition of the SuiBridgeEvent enum, of which each variant is an emitted Event struct defind in the Move Bridge module. We rely on structures in this file to decode the bcs content of the emitted events.

orchestrator

BridgeOrchestrator is the component that:

server

storage

sui_client

sui_syncer

The SuiSyncer module is responsible for synchronizing Events emitted on Sui blockchain from concerned bridge packages.

types

Macros

declare_events

retry_with_max_delay