Module sui_bridge::abi::eth_bridge_vault

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

AddressInsufficientBalance

Custom Error type AddressInsufficientBalance with signature AddressInsufficientBalance(address) and selector 0xcd786059

EthBridgeVault

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

ReentrancyGuardReentrantCall

Custom Error type ReentrancyGuardReentrantCall with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

SafeERC20FailedOperation

Custom Error type SafeERC20FailedOperation with signature SafeERC20FailedOperation(address) and selector 0x5274afe7

TransferERC20Call

Container type for all input parameters for the transferERC20 function with signature transferERC20(address,address,uint256) and selector 0x9db5dbe4

TransferETHCall

Container type for all input parameters for the transferETH function with signature transferETH(address,uint256) and selector 0x7b1a4909

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

WethCall

Container type for all input parameters for the wETH function with signature wETH() and selector 0xf2428621

WethReturn

Container type for all return fields from the wETH function with signature wETH() and selector 0xf2428621

Enums

EthBridgeVaultCalls

Container type for all of the contract’s call

EthBridgeVaultErrors

Container type for all of the contract’s custom errors

Statics

ETHBRIDGEVAULT_ABI

The parsed JSON ABI of the contract.