Module sui_bridge::abi::eth_committee_upgradeable_contract

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

CommitteeCall

Container type for all input parameters for the committee function with signature committee() and selector 0xd864e740

CommitteeReturn

Container type for all return fields from the committee function with signature committee() and selector 0xd864e740

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

EthCommitteeUpgradeableContract

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

InitializedFilter

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

Message

Message(uint8,uint8,uint64,uint8,bytes)

NoncesCall

Container type for all input parameters for the nonces function with signature nonces(uint8) and selector 0x5035bda2

NoncesReturn

Container type for all return fields from the nonces function with signature nonces(uint8) and selector 0x5035bda2

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ReentrancyGuardReentrantCall

Custom Error type ReentrancyGuardReentrantCall with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradeWithSignaturesCall

Container type for all input parameters for the upgradeWithSignatures function with signature upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xfa72a6d0

UpgradedFilter

Enums

EthCommitteeUpgradeableContractCalls

Container type for all of the contract’s call

EthCommitteeUpgradeableContractErrors

Container type for all of the contract’s custom errors

EthCommitteeUpgradeableContractEvents

Container type for all of the contract’s events

Statics

ETHCOMMITTEEUPGRADEABLECONTRACT_ABI

The parsed JSON ABI of the contract.