sui_bridge::abi
Module eth_erc20

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AllowanceCall

Container type for all input parameters for the allowance function with signature allowance(address,address) and selector 0xdd62ed3e

AllowanceReturn

Container type for all return fields from the allowance function with signature allowance(address,address) and selector 0xdd62ed3e

ApprovalFilter

ApproveCall

Container type for all input parameters for the approve function with signature approve(address,uint256) and selector 0x095ea7b3

ApproveReturn

Container type for all return fields from the approve function with signature approve(address,uint256) and selector 0x095ea7b3

BalanceOfCall

Container type for all input parameters for the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BalanceOfReturn

Container type for all return fields from the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BurnCall

Container type for all input parameters for the burn function with signature burn(address,uint256) and selector 0x9dc29fac

DecimalsCall

Container type for all input parameters for the decimals function with signature decimals() and selector 0x313ce567

DecimalsReturn

Container type for all return fields from the decimals function with signature decimals() and selector 0x313ce567

ERC20InsufficientAllowance

Custom Error type ERC20InsufficientAllowance with signature ERC20InsufficientAllowance(address,uint256,uint256) and selector 0xfb8f41b2

ERC20InsufficientBalance

Custom Error type ERC20InsufficientBalance with signature ERC20InsufficientBalance(address,uint256,uint256) and selector 0xe450d38c

ERC20InvalidApprover

Custom Error type ERC20InvalidApprover with signature ERC20InvalidApprover(address) and selector 0xe602df05

ERC20InvalidReceiver

Custom Error type ERC20InvalidReceiver with signature ERC20InvalidReceiver(address) and selector 0xec442f05

ERC20InvalidSender

Custom Error type ERC20InvalidSender with signature ERC20InvalidSender(address) and selector 0x96c6fd1e

ERC20InvalidSpender

Custom Error type ERC20InvalidSpender with signature ERC20InvalidSpender(address) and selector 0x94280d62

EthERC20

MintCall

Container type for all input parameters for the mint function with signature mint(address,uint256) and selector 0x40c10f19

NameCall

Container type for all input parameters for the name function with signature name() and selector 0x06fdde03

NameReturn

Container type for all return fields from the name function with signature name() and selector 0x06fdde03

SymbolCall

Container type for all input parameters for the symbol function with signature symbol() and selector 0x95d89b41

SymbolReturn

Container type for all return fields from the symbol function with signature symbol() and selector 0x95d89b41

TestSkipCall

Container type for all input parameters for the testSkip function with signature testSkip() and selector 0x7cc17ba4

TotalSupplyCall

Container type for all input parameters for the totalSupply function with signature totalSupply() and selector 0x18160ddd

TotalSupplyReturn

Container type for all return fields from the totalSupply function with signature totalSupply() and selector 0x18160ddd

TransferCall

Container type for all input parameters for the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

TransferFilter

TransferFromCall

Container type for all input parameters for the transferFrom function with signature transferFrom(address,address,uint256) and selector 0x23b872dd

TransferFromReturn

Container type for all return fields from the transferFrom function with signature transferFrom(address,address,uint256) and selector 0x23b872dd

TransferReturn

Container type for all return fields from the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

Enums

EthERC20Calls

Container type for all of the contract’s call

EthERC20Errors

Container type for all of the contract’s custom errors

EthERC20Events

Container type for all of the contract’s events

Statics

ETHERC20_ABI

The parsed JSON ABI of the contract.