Module sui_bridge::abi::eth_sui_bridge

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

AddressInsufficientBalance

Custom Error type AddressInsufficientBalance with signature AddressInsufficientBalance(address) and selector 0xcd786059

BridgeERC20Call

Container type for all input parameters for the bridgeERC20 function with signature bridgeERC20(uint8,uint256,bytes,uint8) and selector 0x212afaff

BridgeETHCall

Container type for all input parameters for the bridgeETH function with signature bridgeETH(bytes,uint8) and selector 0x9449ebd2

CommitteeCall

Container type for all input parameters for the committee function with signature committee() and selector 0xd864e740

CommitteeReturn

Container type for all return fields from the committee function with signature committee() and selector 0xd864e740

ContractUpgradedFilter

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

EmergencyOperationFilter

EnforcedPause

Custom Error type EnforcedPause with signature EnforcedPause() and selector 0xd93c0665

EthSuiBridge

ExecuteEmergencyOpWithSignaturesCall

Container type for all input parameters for the executeEmergencyOpWithSignatures function with signature executeEmergencyOpWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xa6f740f6

ExpectedPause

Custom Error type ExpectedPause with signature ExpectedPause() and selector 0x8dfc202b

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address,address,address) and selector 0xc0c53b8b

InitializedFilter

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

IsTransferProcessedCall

Container type for all input parameters for the isTransferProcessed function with signature isTransferProcessed(uint64) and selector 0x5458ea9e

IsTransferProcessedReturn

Container type for all return fields from the isTransferProcessed function with signature isTransferProcessed(uint64) and selector 0x5458ea9e

LimiterCall

Container type for all input parameters for the limiter function with signature limiter() and selector 0x74b87f67

LimiterReturn

Container type for all return fields from the limiter function with signature limiter() and selector 0x74b87f67

Message

Message(uint8,uint8,uint64,uint8,bytes)

NoncesCall

Container type for all input parameters for the nonces function with signature nonces(uint8) and selector 0x5035bda2

NoncesReturn

Container type for all return fields from the nonces function with signature nonces(uint8) and selector 0x5035bda2

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

PausedCall

Container type for all input parameters for the paused function with signature paused() and selector 0x5c975abb

PausedFilter

PausedReturn

Container type for all return fields from the paused function with signature paused() and selector 0x5c975abb

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ReentrancyGuardReentrantCall

Custom Error type ReentrancyGuardReentrantCall with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5

SafeERC20FailedOperation

Custom Error type SafeERC20FailedOperation with signature SafeERC20FailedOperation(address) and selector 0x5274afe7

TokensClaimedFilter

TokensDepositedFilter

TransferBridgedTokensWithSignaturesCall

Container type for all input parameters for the transferBridgedTokensWithSignatures function with signature transferBridgedTokensWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xbeb0d55c

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UnpausedFilter

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradeWithSignaturesCall

Container type for all input parameters for the upgradeWithSignatures function with signature upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xfa72a6d0

UpgradedFilter

VaultCall

Container type for all input parameters for the vault function with signature vault() and selector 0xfbfa77cf

VaultReturn

Container type for all return fields from the vault function with signature vault() and selector 0xfbfa77cf

Enums

EthSuiBridgeCalls

Container type for all of the contract’s call

EthSuiBridgeErrors

Container type for all of the contract’s custom errors

EthSuiBridgeEvents

Container type for all of the contract’s events

Statics

ETHSUIBRIDGE_ABI

The parsed JSON ABI of the contract.