sui_bridge
Module events

This file contains the definition of the SuiBridgeEvent enum, of which each variant is an emitted Event struct defind in the Move Bridge module. We rely on structures in this file to decode the bcs content of the emitted events.

Structs

BlocklistValidatorEvent

CommitteeMemberUrlUpdateEvent

CommitteeUpdate

EmergencyOpEvent

EmittedSuiToEthTokenBridgeV1

MoveBlocklistValidatorEvent

MoveCommitteeMemberUrlUpdateEvent

MoveCommitteeUpdateEvent

MoveNewTokenEvent

MoveTokenDepositedEvent

MoveTokenRegistrationEvent

MoveTokenTransferAlreadyApproved

MoveTokenTransferAlreadyClaimed

MoveTokenTransferApproved

MoveTokenTransferClaimed

MoveTokenTransferLimitExceed

NewTokenEvent

TokenRegistrationEvent

TokenTransferAlreadyApproved

TokenTransferAlreadyClaimed

TokenTransferApproved

TokenTransferClaimed

TokenTransferLimitExceed

UpdateRouteLimitEvent

UpdateTokenPriceEvent

Enums

SuiBridgeEvent

Statics

BlocklistValidatorEvent

CommitteeMemberRegistration

CommitteeMemberUrlUpdateEvent

CommitteeUpdateEvent

EmergencyOpEvent

NewTokenEvent

SuiToEthTokenBridgeV1

TokenRegistrationEvent

TokenTransferAlreadyApproved

TokenTransferAlreadyClaimed

TokenTransferApproved

TokenTransferClaimed

TokenTransferLimitExceed

UpdateRouteLimitEvent

UpdateTokenPriceEvent