Module sui_bridge::utils

Structs

EthBridgeContracts

Functions

examine_key

Read bridge key from a file and print the corresponding information. If is_validator_key is true, the key must be a Secp256k1 key.

generate_bridge_authority_key_and_write_to_file

Generate Bridge Authority key (Secp256k1KeyPair) and write to a file as base64 encoded privkey.

generate_bridge_client_key_and_write_to_file

Generate Bridge Client key (Secp256k1KeyPair or Ed25519KeyPair) and write to a file as base64 encoded flag || privkey.

generate_bridge_node_config_and_write_to_file

Generate Bridge Node Config template and write to a file.

get_committee_voting_power_by_name

Return a mappping from validator name to their bridge voting power. If a validator is not in the Sui committee, we will use its base URL as the name.

get_eth_contract_addresses

Given the address of SuiBridge Proxy, return the addresses of the committee, limiter, vault, and config.

get_eth_contracts

Given the address of SuiBridge Proxy, return the contracts of the committee, limiter, vault, and config.

get_eth_signer_client

get_validator_names_by_pub_keys

Return a mappping from validator pub keys to their names. If a validator is not in the Sui committee, we will use its base URL as the name.

publish_and_register_coins_return_add_coins_on_sui_action

wait_for_server_to_be_up

Type Aliases

EthSigner